Refactor code to use recoil for global state
Wrote a custom hook to call the grant list data and update the recoil atom. This is then passed into the grant page, and the subpage/grant data grid. From there the currently clicked on item is then saved into the global state in its own variable. This should make deleting/editing a grant simple and not require us to pass data between components. From now on we should try to design things decoupled as possible.